requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
clickhouse-connect==0.7.7
clickhouse-driver==0.2.7
docker==7.0.0
Faker==24.7.1
idna==3.6
lz4==4.3.3
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandahouse==0.2.7
pandas==2.2.1
pulsar-client==3.4.0
pycountry==23.12.11
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
requests==2.31.0
six==1.16.0
toolz==0.12.1
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
wonderwords==2.2.0
zstandard==0.22.0
